Refactor the package of ResourcePDX
Authored-by: Jianxia Chen <jchen21@apache.org>
