+++ geode-connectors/src/distributedTest/resources/org/apache/geode/connectors/jdbc/internal/cli/ResourcePDX.java
@@ -12,7 +12,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.apache.geode.internal;
+package org.apache.geode.connectors.jdbc.internal.cli;
 
 import org.apache.geode.pdx.PdxReader;
 import org.apache.geode.pdx.PdxSerializable;
